fix: install Python 3.9 from source
We need to support Ubuntu 16.04 as the kong-2.0.5 image is based on it.  Thus,
we cannot rely on a sufficiently current httpie or Python version to be
available in the distribution repository.  We also can't use the deadsnakes
PPA as that does not support Ubuntu 16.04 anymore.  We're thus installing
Python 3.9 from source.