feat: allow building pongo image from specially tagged test image
This change adds the possibility to specify a temporary test image
as KONG_IMAGE, similar to using the 'dev' and 'dev-ee' pseudo versions.
The image name contains the git tag from which it was built.  The
added logic will determine KONG_VERSION and whether we're dealing
with CE or EE from the image name