fix(shell): change prefix for health checks
Since Pongo sets the prefix to use to `/kong-plugin/servroot/` the
healthchecks would always fail because they would check in the
default location. This fixes them.

Health is ok, if Kong is running in the container.
So in case of a `pongo shell` it is unhealthy until Kong is started.